+++ scaladia-container/src/main/scala/com/github/giiita/container/Indexer.scala
