+++ scaladia-container/src/main/scala/com/github/giiita/injector/AutoInject.scala
@@ -1,5 +1,5 @@
 package com.github.giiita.injector
 
-trait AutoInject[T] extends Injector {
+trait AutoInject[T] extends AutoInjectable with Injector {
   com.github.giiita.container.Container.flush(this)
 }
