--- scaladia-container/src/main/scala/com/github/giiita/injector/MacroTest.scala
@@ -1,11 +1,0 @@
-package com.github.giiita.injector
-
-import com.github.giiita.injector.testpack.TTT.MOMO
-
-object MacroTest extends App with Injector {
-  println(
-    inject[MOMO].provide
-  )
-
-
-}
