+++ scaladia-container/src/main/scala/com/github/giiita/testpack/deeper/MMMMMM.scala
@@ -1,0 +1,5 @@
+package com.github.giiita.testpack.deeper
+
+class MMMMMM {
+
+}
