+++ scaladia-macro/src/main/scala/com/github/giiita/injector/AutoInjectable.scala
@@ -1,0 +1,3 @@
+package com.github.giiita.injector
+
+trait AutoInjectable
