--- scaladia-macro/src/main/scala/com/github/giiita/injector/testpack/deeper/MMMMMM.scala
@@ -1,5 +1,0 @@
-package com.github.giiita.injector.testpack.deeper
-
-class MMMMMM {
-
-}
